Refactor Permission controls; add a vertical list for a single model; use that for convention-level permissions
